Rename aux source to auxi source
The 'aux' file name is a special name on Windows operating systems, so
Windows has difficulty when 'aux' is used as a name for regular files.
