+++ src/auxi.h
