+++ tests/test-aux8-pngwrite.c
@@ -9,7 +9,7 @@
  * png writer test program
  */
 
-#include "../src/aux.h"
+#include "../src/auxi.h"
 #include <stdio.h>
 #include <errno.h>
 #include <string.h>
